Updated Inventory/ Player classes. Chop Wood does not crash any more

+++ src/edu/ycp/cs320/IslandAdventure/controller/ActionController.java
@@ -1,7 +1,5 @@
 package edu.ycp.cs320.IslandAdventure.controller;
 
-import edu.ycp.cs320.IslandAdventure.model.*;
-
 public class ActionController 
 {
 	InventoryController inventoryController = new InventoryController();
@@ -10,7 +8,7 @@ public class ActionController
 	{
 		if (action.equals("Chop Wood") || action.equals("chop wood")) 
 		{
-			inventoryController.changeWoodAmount((double) 10.0);
+			inventoryController.changeWoodAmount(10);
 		}
 	}
 }
